pyproject.toml
[tool.poetry]
name = "gradcafe-updater"
version = "1.0.0"
description = "The GRADECAFE updater, automate the notification."
authors = ["Louie Lu <git@louie.lu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
lxml = "^4.9.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
